修复 default 对象为 undefined 的问 题
